requirements.txt
appdirs
atomicwrites
attrs
black
certifi
cffi
chardet
click
colorama
coverage
cryptography
Django
django-js-asset
django-mptt
factory-boy
Faker
gunicorn
idna
inflection
iniconfig
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
Pillow==8.1.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-django==4.1.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3